modified:   Templates/loq/interval_detail.html 	modified:   loq/graph_view.py 	modified:   loq/graph_view.pyc 	modified:   loq/interval_detail_view.py 	modified:   loq/interval_detail_view.pyc 	new file:   loq/migrations/0030_auto__add_field_interval_mid_point.py 	new file:   loq/migrations/0030_auto__add_field_interval_mid_point.pyc 	modified:   loq/models.py 	modified:   loq/models.pyc

+++ loq/graph_view.py
@@ -51,7 +51,7 @@ def plot_graph(name,libraries,chart_type,normal,style):
             data = i['read_length__count']*i['read_counts']
             title = 'Reads Counts'
         else: 
-            data = (100*i['read_length__count']*i['read_counts'])/d[i['lib']]
+            data = (100*float(i['read_length__count']*i['read_counts']))/float(d[i['lib']])
             title ='Read counts per miRNA mapped'
             
         df[lb][lh] += data
